+++ elements/drawtools/stories/index.js
@@ -7,8 +7,9 @@ export { default as MultiPolygonStory } from "./multi-polygon"; // Story showcas
 export { default as ModifyFeaturesStory } from "./modify-features"; // Story illustrating different draw types.
 export { default as MultiPolygonWithListStory } from "./multi-polygon-list"; // Story demonstrating polygons with lists.
 export { default as ImportFeaturesWithEditorStory } from "./import-features"; //
-export { default as UnstyledStory } from "./unstyled"; // Story demonstrating unstyled variant of element.
-export { default as CSSVariableOverrideStory } from "./css-variable-override"; // Story demonstrating css override
 export { default as SelectFeatureStory } from "./select-feature"; // Story demonstrating feature selection
 export { default as MultiFeaturesSelectStory } from "./multi-feature-select"; // Story demonstrating multi feature selection with list
+export { default as FeaturesProjectionStory } from "./features-projection"; // Story showcasing emitting drawn features in different projections
+export { default as CSSVariableOverrideStory } from "./css-variable-override"; // Story demonstrating css override
 export { default as DistanceMeasurementStory } from "./distance-measurement"; // Story demonstrating distance measurements.
+export { default as UnstyledStory } from "./unstyled"; // Story demonstrating unstyled variant of element.