+++ elements/itemfilter/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [1.5.0](https://github.com/EOX-A/EOxElements/compare/itemfilter-v1.4.1...itemfilter-v1.5.0) (2024-11-07)
+
+
+### Features
+
+* Build and include element's type declarations ([#1353](https://github.com/EOX-A/EOxElements/issues/1353)) ([f17166e](https://github.com/EOX-A/EOxElements/commit/f17166e292ce546a2ff45433a05248330eb63713))
+
+## [1.4.1](https://github.com/EOX-A/EOxElements/compare/itemfilter-v1.4.0...itemfilter-v1.4.1) (2024-11-04)
+
+
+### Bug Fixes
+
+* Fall back to filter key if no title is provided (chip display) ([#1328](https://github.com/EOX-A/EOxElements/issues/1328)) ([d1aaf2d](https://github.com/EOX-A/EOxElements/commit/d1aaf2d4c7cfbf797d57cc70b8c0942379c06b0a))
+
 ## [1.4.0](https://github.com/EOX-A/EOxElements/compare/itemfilter-v1.3.1...itemfilter-v1.4.0) (2024-10-17)